requirements.txt
pandas==2.3.2
nltk==3.9.1
openai==0.27.0
python-dotenv==1.1.1
matplotlib==3.10.6
seaborn==0.13.2
errant==3.0.0
fastapi==0.117.1